Adds support for outlets
Outlets allow for alternate paths from the main path. I'm not sure
if I'm going to keep the name outlet; fork is probably more
appropriate.

The problem with this implementation of forks is that all of the
upstream actions are repeated every time a fork is used. It would
be nice to cache a version of the upstream data for the fork to use
instead.