Outlets receive a copy of the data at that point
When an outlet is created, the Path is executed with the current
steps it has on the stack. This has several benefits:

* We don't have to repeat any of the previous transformations
* The interface is much cleaner

The cost is memory, because we're basically making a copy of the
data for the outlet to use while the parent path is still being
built. In other words, it's not lazy.