Adds dependencies and execution contexts
I wanted to allow a way to define dependencies that could be used
during the realization of a path. So, we give the dependency a name
and block that resolves the dependency. When the path is being
realized with data, any blocks will be executed within the context.
Also, all dependencies are added to the context as instance methods
with their name for easy access. For example:

```
DataPath::Path.new do
  depends_on(:external_data) do
    get_external_data_from_a_weird_source()
  end

  calc(:external_id) do |data|
    external_row = external_data.find do |other|
      other[:internal_id] == data[:id]
    end

    external_row[:external_id]
  end
end
```