Data can be exploded
Sometimes rows need to be added to a pipeline. The explode command
will add the result array of the step into the path such that each
member of the array is treated as a new piece of data.

This change required some major changes in the way that the path
is executed. Previously the path would have 0 or 1 results for a
single piece of data. Now there can be 0, 1 or more results. To
handle this gracefully I introduced step handlers.

Step handlers have an input, and an output. Source data is pushed
into the handler, and the output can contain 0 (filtered), 1 or
more results. The handler is responsibly for determining the output.