Copies dependencies into forks
When forking a path, the dependencies from the parent were
unavailable in the forked path.

Now when a path is forked, the parent's dependencies are copied
into the fork.

Also, in order for the context to be passed to the parent's
`#call` method, we had to check for the absolut value of the
arity, since the context is an optional argument.