Makes forks available on the realization object
Instead of going back to the path to realize a fork, the same data
can be processed by the fork from the original realization.

This commit also tightens access to the forks on the path. Instead
of giving access to the whole forks hash, forks can only be accessed
by the #forks method.