Preserves data that will be forked
Before this commit, when a fork transformed data it would reprocess
the data from the original source. With this change, a copy of the
data is preserved right before the fork so that it can use that
copy and apply only the changes in the fork on that data instead
of the original source.