Implements an #each_slice method on the transformation
The default #each_slice method would always return an array. This
was problematic because the forked data wasn't available on the
slice.

Now #each_slice returns a new transformation with just the data
for that slice, including forks.

The side effect of doing this as that the transformation now
requires an full Enumerable, not just an object that responds
to #each. This is a fair trade-off because most objects that
respond to #each should likely be Enumerable anyway.