Memoizes the results of #each_slice
I'm not totally sure I like this, it could cause memory bloat
if each_slice is called with different sizes.

More than likely each_slice will only be called once, so I might
take this out in the future. For now, it is at least consistent
with the memoization of #each