Specs for all of the readme examples
To ensure that the README examples are correct and accurate, I
implemented a spec for each one. During this process, I found
several bugs (added in previous commits). Previous commits
fixed some of these bugs, and now the readme specs are complete.

Due the to bugs, this commit will also bump the version.