Adds configurable error handling
There are two modes for handling errors.

`raise_errors = true` means that errors will be raised if they
occur while executing a step. This will abort the transformation.

`raise_errors = false` errors will not be raised. Instead, the data
will be skipped and the error details about the error will be added
to the transformation.

The default value for raise_errors is true.