Improves error handling with StepExecutionError
When an error occurs during transformation while executing a step,
a StepExecutionError will be raised. This will provide the cause
of the error, the input data, the data at the time the error
occurred, and the step which caused the error.

Whether or not the error is thrown or added to the transformation's
errors collection it will always be an instance of the
StepExecutionError.