Refactors the Handler classes into Steps
Composition over inheritance! Now there is a single StepHandler
class that is responsible for the movement of data. It will
receive one an instance of a Step class when it is instanciated.

The step will be called and should yield data that needs to be
pushed out by the StepHandler.

Overall, this is a better a design as it separates the
responsibilities of handling the input and output of the data from
the actual step that transforms the data.